Refactor models.rs into structure matching spec
