+++ cyclonedx-bom/src/models/metadata.rs
@@ -1,0 +1,36 @@
+/*
+ * This file is part of CycloneDX Rust Cargo.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use crate::external_models::date_time::DateTime;
+use crate::models::component::Component;
+use crate::models::license::Licenses;
+use crate::models::organization::{OrganizationalContact, OrganizationalEntity};
+use crate::models::property::Properties;
+use crate::models::tool::Tools;
+
+#[derive(Debug, PartialEq)]
+pub struct Metadata {
+    pub timestamp: Option<DateTime>,
+    pub tools: Option<Tools>,
+    pub authors: Option<Vec<OrganizationalContact>>,
+    pub component: Option<Component>,
+    pub manufacture: Option<OrganizationalEntity>,
+    pub supplier: Option<OrganizationalEntity>,
+    pub licenses: Option<Licenses>,
+    pub properties: Option<Properties>,
+}
