+++ cyclonedx-bom/src/models/property.rs
@@ -1,0 +1,28 @@
+/*
+ * This file is part of CycloneDX Rust Cargo.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use crate::external_models::normalized_string::NormalizedString;
+
+#[derive(Debug, PartialEq)]
+pub struct Properties(pub(crate) Vec<Property>);
+
+#[derive(Debug, PartialEq)]
+pub struct Property {
+    pub name: String,
+    pub value: NormalizedString,
+}
