+++ cyclonedx-bom/src/models/tool.rs
@@ -1,0 +1,31 @@
+/*
+ * This file is part of CycloneDX Rust Cargo.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use crate::external_models::normalized_string::NormalizedString;
+use crate::models::hash::Hashes;
+
+#[derive(Debug, PartialEq)]
+pub struct Tool {
+    pub vendor: Option<NormalizedString>,
+    pub name: Option<NormalizedString>,
+    pub version: Option<NormalizedString>,
+    pub hashes: Option<Hashes>,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct Tools(pub Vec<Tool>);
